requirements.txt
imageio==2.26.0
matplotlib==3.8.3
numpy==1.23.1
open3d==0.18.0
opencv_contrib_python==4.9.0.80
opencv_python==4.7.0.72
opencv_python_headless==4.9.0.80
pandas==1.5.3
Pillow==9.4.0
Pillow==10.3.0
pointnet2_ops==3.0.0
PyYAML==6.0.1
PyYAML==6.0.1
scikit_learn==1.2.1
scipy==1.9.1
tifffile==2023.4.12
timm==1.0.3
torch==1.13.1
torchvision==0.14.1
tqdm==4.65.0
wandb==0.15.5
